+++ graphile/graphile-sql-expression-validator/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.9.3](https://github.com/constructive-io/constructive/compare/graphile-sql-expression-validator@3.9.2...graphile-sql-expression-validator@3.9.3) (2026-07-31)
+
+**Note:** Version bump only for package graphile-sql-expression-validator
+
 ## [3.9.2](https://github.com/constructive-io/constructive/compare/graphile-sql-expression-validator@3.9.1...graphile-sql-expression-validator@3.9.2) (2026-07-31)
 
 **Note:** Version bump only for package graphile-sql-expression-validator