+++ graphql/explorer/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.10.6](https://github.com/constructive-io/constructive/compare/@constructive-io/graphql-explorer@5.10.5...@constructive-io/graphql-explorer@5.10.6) (2026-07-31)
+
+**Note:** Version bump only for package @constructive-io/graphql-explorer
+
 ## [5.10.5](https://github.com/constructive-io/constructive/compare/@constructive-io/graphql-explorer@5.10.4...@constructive-io/graphql-explorer@5.10.5) (2026-07-31)
 
 **Note:** Version bump only for package @constructive-io/graphql-explorer