+++ graphql/playwright-test/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.13.6](https://github.com/constructive-io/constructive/compare/@constructive-io/playwright-test@3.13.5...@constructive-io/playwright-test@3.13.6) (2026-07-31)
+
+**Note:** Version bump only for package @constructive-io/playwright-test
+
 ## [3.13.5](https://github.com/constructive-io/constructive/compare/@constructive-io/playwright-test@3.13.4...@constructive-io/playwright-test@3.13.5) (2026-07-31)
 
 **Note:** Version bump only for package @constructive-io/playwright-test