+++ packages/csv-to-pg/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.9.3](https://github.com/constructive-io/constructive/compare/csv-to-pg@4.9.2...csv-to-pg@4.9.3) (2026-07-31)
+
+**Note:** Version bump only for package csv-to-pg
+
 ## [4.9.2](https://github.com/constructive-io/constructive/compare/csv-to-pg@4.9.1...csv-to-pg@4.9.2) (2026-07-31)
 
 **Note:** Version bump only for package csv-to-pg