+++ packages/safegres/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.15.1](https://github.com/constructive-io/constructive/compare/safegres@1.15.0...safegres@1.15.1) (2026-07-31)
+
+**Note:** Version bump only for package safegres
+
 # [1.15.0](https://github.com/constructive-io/constructive/compare/safegres@1.14.2...safegres@1.15.0) (2026-07-31)
 
 ### Features